backend/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
PyJWT
python-multipart